update2
Antes de incluir uma lista de funcionários

+++ Projeto_Final_Programacao_2/Sitema_Biblioteca/src/br/com/unit/sistemabiblioteca/business/controller/EmprestimoController.java
@@ -10,7 +10,7 @@ import br.com.unit.sistemabiblioteca.dao.EmprestimoDao;
 
 public class EmprestimoController {
 	EmprestimoDao emprestimoDao;
-	
+
 	public EmprestimoController() {
 		emprestimoDao = new EmprestimoDao();
 	}
@@ -24,35 +24,52 @@ public class EmprestimoController {
 		emprestimo.setFuncionario(funcionario);
 		emprestimo.setLeitor(leitor);
 		emprestimo.setLivro(livro);
-		
+
 		emprestimoDao.inserir(emprestimo);
-		
+
 	}
-	
+
 	public Emprestimo consultar(long cpf) {
 		return emprestimoDao.consultar(cpf);
+	}
+	
+	public Emprestimo consultar(long cpf, long codigo) {
+		return emprestimoDao.consultar(cpf,codigo);
 	}
 	
 	public List<Emprestimo> retornarEmprestimosLeitor(long cpf) {
 		return emprestimoDao.retornarEmprestimosLeitor(cpf);
 	}
-	
-	public List<Emprestimo> retornarTodos(){
+
+	public List<Emprestimo> retornarTodos() {
 		return emprestimoDao.retornarTodos();
 	}
 
-	public void removerEmprestimoLivro(long cpf, long codigo){
-		emprestimoDao.removerEmprestimoLivro(cpf, codigo);
+	public void removerEmprestimoLivro(long cpf, long codigo) throws Exception {
+		Emprestimo emprestimo = consultar(cpf, codigo);
+		if (existe(emprestimo)) {
+			emprestimoDao.removerEmprestimoLivro(cpf, codigo);
+		} else {
+			throw new Exception("Emprétimo não existe!");
+		}
 	}
 
-	public void renovar(Leitor leitor, Livro livro, Funcionario funcionario) {
+	public void renovar(Leitor leitor, Livro livro, Funcionario funcionario) throws Exception {
 		Emprestimo emprestimo = new Emprestimo();
 		emprestimo.setLeitor(leitor);
 		emprestimo.setFuncionario(funcionario);
 		emprestimo.setLivro(livro);
-		
-		emprestimoDao.renovar(emprestimo);
-		
+		if (existe(emprestimo)) {
+			emprestimoDao.renovar(emprestimo);
+		}
 	}
-	
+
+	public boolean existe(Emprestimo emprestimo) throws Exception {
+		if (emprestimoDao.existe(emprestimo)) {
+			return true;
+		} else {
+			throw new Exception("Empréstimo não existe!");
+		}
+	}
+
 }
